requirements.txt
flask==1.1.2
py-healthcheck==1.10.1
celery==4.4.3
boto==2.49.0
boto3==1.13.22
botocore==1.16.23